requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-WTF==1.2.1
WTForms==3.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.0

# Google API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.0

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3
Pillow==10.1.0
pypdf==3.17.1

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1

# LLM and AI
openai==1.3.7
anthropic==0.7.4
tiktoken==0.5.2

# NLP and text processing
spacy==3.7.2
pytesseract==0.3.10

# Citation parsing
pyparsing==3.1.1

# Data handling
pandas==2.1.3
openpyxl==3.1.2

# Environment and config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilities
click==8.1.7
tqdm==4.66.1
pyyaml==6.0.1
python-dateutil==2.8.2

# Logging and monitoring
python-json-logger==2.0.7

# Security
cryptography==41.0.7
PyJWT==2.8.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production server (for local dev)
gunicorn==21.2.0
